Add files via upload
Business Problem Understanding
The shipment process constitutes of several important components. The last mile
delivery is carried out by the delivery boys, which are working on the company’s
payroll. These delivery personnel are many times over-loaded with the task of the
deliveries, as company cannot afford to lose the customer because of bad
impression at time of delivery. Due to the heavy workload, these delivery boys are not
able to perform to their optimum limits on every business day, which results into their
absenteeism from the work, which in turn increases the load on the other delivery
personnel’s.
You are the data scientist appointed to take a closer look at the absenteeism records
of delivery personnel’s and
a) Understand the absenteeism at work
b) Suggest a model that can help to determine the absenteeism hours for an
employee
c) Suggest ways by which delivery performance can be optimized.

Data preparation and Exploratory Data Analysis
You are supposed to make utilizations of all the appropriate data pre-processing
techniques on the given data set. If required, make appropriate assumptions and
make it explicitly known while using them in the code or in the presentation. You are
required to identify the key factors that influences the absenteeism at the work. Make
appropriate selection of the attributes with sound justification for the same. The data
set allows for several new combinations of attributes and attribute exclusions, or the
modification of the attribute type (categorical, integer, or real) depending on the
purpose of the research.
You are supposed to make use of Python programming language and its libraries to
work on this analysis effort.